requirements.txt
aiohttp==3.9.1
websockets==12.0
redis==5.0.1
python-dotenv==1.0.0